refactor: clean code

+++ src/fs/rename.js
@@ -1,8 +1,7 @@
 import * as fs from 'node:fs/promises';
 import { fileURLToPath } from 'node:url';
-import { errorHandler } from '../error-handler.js';
-import { ERROR_CODES } from '../consts.js';
-import { getPath } from '../get-path.js';
+import { errorHandler, getPath, doesFileExist } from '../utils.js';
+import { errorMessage, ERROR_CODES } from '../consts.js';
 
 const wrongFileName = 'wrongFilename.txt';
 const properFileName = 'properFilename.md';
@@ -11,15 +10,13 @@ const targetFilePath = getPath(fileURLToPath(import.meta.url), properFileName);
 
 const rename = async (oldFilePath, newFilePath) => {
     try {
-        await fs.stat(oldFilePath);
-    } catch (error) {
-        errorHandler(error, ERROR_CODES.noSuchFileOrDirectory);
-    }
-
-    try {
+        if (await doesFileExist(newFilePath)) {
+            throw Error(errorMessage);
+        }
+        
         await fs.rename(oldFilePath, newFilePath);
     } catch (error) {
-        errorHandler(error, ERROR_CODES.fileExists);
+        errorHandler(error, ERROR_CODES.noSuchFileOrDirectory);
     }
 };
